+++ src/app/api/auth/[auth0]/route.ts
@@ -2,6 +2,7 @@ import { handleAuth, handleLogin } from "@auth0/nextjs-auth0";
 
 export const GET = handleAuth({
   login: handleLogin({
+    returnTo: "/profile",
     authorizationParams: {
       audience: process.env.AUTH0_AUDIENCE,
       scope: "openid profile email",
